border-collapse width issue
https://github.com/jmosbech/StickyTableHeaders/pull/17

When the table has border-collapse use a different method of calculating
width:

Uses document.defaultView.getComputedStyle to calculate the width
instead of jQuery.width.

This results in a pixel perfect result for me in Chrome and Firefox.
IE8 doesn't support defaultView. I have created a fallback that gets the
outerWidth and subtracts the padding and one lot of the border. IE9+
is supposed to support defaultView and should work too, but I have
not tested it.