add optional parameter for additional scrollable areas
When using "table-responsive" there are actually two scrollable areas. The window and the table-responsive-div. Bot need an event listener on "scroll". Since StickyTableHeaders should not know about bootstrap specific things and it could be similar with other UI frameworks this solution is more generic where the user can add scrollable areas that should be observed on his own.